add reset rollscount when a rule is selected
